Added the Málco logo, email address and other various minor changes
Added the Málco logo. Finnbogi gave me a logo pdf with 3 variations of the logo (with slogan and phone number). I cropped it and exported it to a PNG with a width of 2000px. Size 152KB.

Made the background more white. I'm still using the same base blue colour but just made it more transparent. The background is now mostly blue with a hint of grey and blue which I think looks slightly nicer and creates a better contrast with the black font (easier to read).