Changed font to Roboto light, use icons for phone & email, etc
Updated the meta description to include more information that just the name of the company. This is the description text that search engines sometime use.

Used Roboto Light instead of Open Sans. Roboto Light is more in line with the slogan & phone number font the logo designer choose: Saral Devangari Light. Increased the font size while I was at it.

Replaced "Sími: " and "Póstur: " with Font Awesome icons of a phone and an envelope.

Added the original PDF logo export from the designer to the root of the project. For safe keeping.